Moving CDC tests to functional package
