+++ hudi-spark-datasource/hudi-spark/src/test/scala/org/apache/hudi/functional/TestCDCDataFrameSuite.scala
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.hudi.functional.cdc
+package org.apache.hudi.functional
 
 import org.apache.avro.generic.GenericRecord
 import org.apache.hudi.DataSourceWriteOptions
