+++ hudi-spark-datasource/hudi-spark/src/test/scala/org/apache/hudi/functional/TestCDCStreamingSuite.scala
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.hudi.functional.cdc
+package org.apache.hudi.functional
 
 import org.apache.hudi.common.table.cdc.HoodieCDCSupplementalLoggingMode
 import org.apache.hudi.common.table.{HoodieTableConfig, HoodieTableMetaClient}
